+++ src/ll/misc.py
@@ -429,12 +429,14 @@ def itersplitat(string, positions):
 	"""
 	Split :var:`string` at the positions specified in :var:`positions`.
 
-	This is a generator. For example::
+	For example::
 
 		>>> from ll import misc
 		>>> import datetime
 		>>> datetime.datetime(*map(int, misc.itersplitat("20090609172345", (4, 6, 8, 10, 12))))
 		datetime.datetime(2009, 6, 9, 17, 23, 45)
+
+	This is a generator.
 	"""
 	curpos = 0
 	for pos in positions: