+++ NEWS.rst
@@ -8,8 +8,8 @@ Changes in 3.7 (released ??/??/2009)
 
 *	:mod:`ll.orasql` and :mod:`ll.nightshade` are now part of the distribution.
 
-*	:mod:`ll.make` has a new Action class: :class:`ObjectAction` that simply
-	returns an existing object.
+*	:mod:`ll.make` has a new Action class: :class:`ObjectAction` simply returns
+	an existing object.
 
 *	The following classes have been removed from :mod:`ll.make`:
 	:class:`EncodeAction`, :class:`DecodeAction`, :class:`EvalAction`,